Traits for Actions class
