+++ system/engine/action.php
@@ -8,40 +8,14 @@
 
 namespace Phacil\Framework;
 
-use Phacil\Framework\Interfaces\Action as ActionInterface;
+use \Phacil\Framework\Interfaces\Action as ActionInterface;
+use \Phacil\Framework\Traits\Action as ActionTrait;
 
 /** @package Phacil\Framework */
 final class Action implements ActionInterface {
-	/**
-	 * 
-	 * @var string
-	 */
-	protected $file;
-	
-	/**
-	 * 
-	 * @var string
-	 */
-	protected $class;
-	
-	/**
-	 * 
-	 * @var string
-	 */
-	protected $method;
-	
-	/**
-	 * 
-	 * @var array
-	 */
-	protected $args = array();
 
-	/**
-	 * 
-	 * @var (string[]|string|null)[]
-	 */
-	private $classAlt = [];
-
+	use ActionTrait;
+	
 	/**
 	 * @param string $route 
 	 * @param array $args 
@@ -145,79 +119,14 @@ final class Action implements ActionInterface {
 
 	}
 	
-	/** @return string  */
-	public function getFile():string {
-		return $this->file;
-	}
 	
-	/** @return string  */
-	public function getClass():string {
-		return $this->class;
-	}
-
-	private function mountClass(string $namespace, string $class) {
-		return (defined('NAMESPACE_PREFIX') ? NAMESPACE_PREFIX."\\" : "").str_replace("/", "\\", $namespace)."Controller\\".$class;
-	}
-
-	/**
-	 * 
-	 * @param string $class 
-	 * @return $this 
-	 */
-	public function setClass($class) {
-		$this->class = $class;
-		return $this;
-	}
-
-	/** @return array  */
-	public function getClassAlt():array {
-		return $this->classAlt;
-	}
-	
-	/** @return string  */
-	public function getMethod():string {
-		return $this->method;
-	}
-	
-	/** @return array  */
-	public function getArgs():array {
-		return $this->args;
-	}
 }
 
 
 /** @package Phacil\Framework */
 final class ActionSystem implements ActionInterface {
 	
-	/**
-	 * 
-	 * @var string
-	 */
-	protected $file;
-	
-	/**
-	 * 
-	 * @var string
-	 */
-	protected $class;
-	
-	/**
-	 * 
-	 * @var string
-	 */
-	protected $method;
-	
-	/**
-	 * 
-	 * @var array
-	 */
-	protected $args = array();
-
-	/**
-	 * 
-	 * @var (string[]|string|null)[]
-	 */
-	private $classAlt = [];
+	use ActionTrait;
 
 	/**
 	 * @param string $route 
@@ -269,38 +178,4 @@ final class ActionSystem implements ActionInterface {
 		}
 	}
 	
-	/** @return string  */
-	public function getFile():string {
-		return $this->file;
-	}
-	
-	/** @return string  */
-	public function getClass():string {
-		return $this->class;
-	}
-
-	/**
-	 * 
-	 * @param string $class 
-	 * @return $this 
-	 */
-	public function setClass($class) {
-		$this->class = $class;
-		return $this;
-	}
-	
-	/** @return array  */
-	public function getClassAlt():array {
-		return $this->classAlt;
-	}
-
-	/** @return string  */
-	public function getMethod():string {
-		return $this->method;
-	}
-	
-	/** @return array  */
-	public function getArgs():array {
-		return $this->args;
-	}
 }
